Added Crosshair Change, harvest resource
-Crosshair change
Changes crosshair to hand when detecting a game object that is a
resource

-Harvest Resource (In progress)
player can now use the interact key (E) to harvest a resource. This
destroys the game object and creates an appropriate instance of
resource.

Needs to be shown in ui and added to characterStats or new class for
inventory.